save result in local
